refactor(indexer): flatten perps_trades cache into one file
`perps_trades/cache.rs` was the only file in its sub-directory and its
parent `perps_trades.rs` was a one-line `pub mod cache;`. Fold the cache
into the parent file so `PerpsTradeCache` lives at
`indexer_sql::write::perps_trades::PerpsTradeCache`.
