+++ dango/indexer/httpd/src/context.rs
@@ -1,7 +1,7 @@
 use {
     indexer_sql::{
         entity::perps_trade::PerpsTrade, pubsub::PubSub,
-        write::perps_trades::cache::PerpsTradeCache,
+        write::perps_trades::PerpsTradeCache,
     },
     sea_orm::DatabaseConnection,
     std::sync::Arc,
